% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv.teal-package.R
\docType{package}
\name{dv.teal-package}
\alias{dv.teal}
\alias{dv.teal-package}
\title{dv.teal: DaVinci's teal wrapper package}
\description{
The dv.teal package provides a wrapper function to include teal modules within the DaVinci framework.
}
\author{
\strong{Maintainer}: Korbinian Matthias \email{korbinian.matthias@boehringer-ingelheim.com}

Other contributors:
\itemize{
  \item Boehringer-Ingelheim Pharma GmbH & Co.KG [copyright holder, funder]
}

}
\keyword{internal}
