% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_mod_teal.R
\name{mock_mod_teal}
\alias{mock_mod_teal}
\title{Mock app with one teal module (tm_t_events).}
\usage{
mock_mod_teal()
}
\description{
\code{mock_mod_teal} launches a mock app with the tm_t_events teal module by means of the module manager
(dv.manager).
}
