% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_mod_teal.R
\name{mock_mod_teal_multi}
\alias{mock_mod_teal_multi}
\title{Mock app with multiple teal modules (m_variable_browser, tm_outliers, tm_t_summary).}
\usage{
mock_mod_teal_multi()
}
\description{
\code{mock_mod_teal_multi} launches a mock app with the tm_variable_browser, tm_outliers, and tm_t_summary
teal modules by means of the module manager (dv.manager).
}
