% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pharmaversesdtm.R
\docType{data}
\name{pharmaversesdtm}
\alias{pharmaversesdtm}
\alias{dm}
\alias{ae}
\title{\code{CDISC} Data for Examples and tests from the \code{pharmaversesdtm} package}
\format{
\code{rds} (\code{data.frame})

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 306 rows and 25 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 35 columns.
}
\usage{
dm

ae
}
\description{
\code{CDISC} Data for Examples and tests from the \code{pharmaversesdtm} package
}
\section{Functions}{
\itemize{
\item \code{dm}: \code{dm} data

\item \code{ae}: \code{ae} data

}}
\keyword{datasets}
