% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv.teal-package.R
\docType{package}
\name{dv.teal-package}
\alias{dv.teal}
\alias{dv.teal-package}
\title{dv.teal: DaVinci's teal wrapper package}
\description{
The dv.teal package provides wrapper functions to include teal modules within the DaVinci framework.
}
\author{
\strong{Maintainer}: Korbinian Matthias \email{korbinian.matthias@boehringer-ingelheim.com}

Authors:
\itemize{
  \item Ming Yang \email{ming.yang.ext@boehringer-ingelheim.com}
}

}
\keyword{internal}
